requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
prophet==1.1.5
statsmodels==0.14.1
scikit-learn==1.4.1.post1
tensorflow-cpu==2.15.0